+++ saw-remote-api/python/tests/saw/test_provers.py
@@ -20,6 +20,7 @@ class ProverTest(unittest.TestCase):
         simple_thm = cry('\(x:[8]) -> x != x+1')
         #self.assertTrue(saw.prove(simple_thm, ProofScript([yices([])])).is_valid())
         self.assertTrue(saw.prove(simple_thm, ProofScript([z3([])])).is_valid())
+        self.assertTrue(saw.prove(simple_thm, ProofScript([Quickcheck(100)])).is_valid())
 
         self.assertTrue(saw.prove(simple_thm, ProofScript([Admit()])).is_valid())
         self.assertTrue(saw.prove(cry('True'), ProofScript([Trivial()])).is_valid())